Simplify and unify the layout for both desktop and mobile
