+++ projects/packages/publicize/_inc/components/customize-and-preview/customization-section/styles.module.scss
@@ -1,6 +1,15 @@
+@use "@automattic/jetpack-base-styles/gutenberg-base-styles" as gb;
+
 .customization-section {
 	padding: 1rem;
 	display: flex;
 	flex-direction: column;
 	gap: 1.5rem;
+
+	@include gb.break-medium() {
+
+		&[data-variant="global"] {
+			border-inline-end: 1px solid gb.$gray-200;
+		}
+	}
 }
