+++ projects/packages/publicize/_inc/components/customize-and-preview/customization-toggle/styles.module.scss
@@ -1,0 +1,5 @@
+.customization-toggle {
+	max-height: 2rem;
+	max-width: 350px;
+	margin: 0.5rem auto;
+}
