+++ projects/packages/publicize/_inc/components/customize-and-preview/styles.module.scss
@@ -1,18 +1,20 @@
-@use "@automattic/jetpack-base-styles/gutenberg-base-styles" as gb;
-
 .customize-and-preview {
 	height: 100%;
 	display: grid;
-	grid-template-rows: auto 1fr;
+	grid-template-rows: 1fr;
+
+	&[data-plan="paid"] {
+		grid-template-rows: auto 1fr;
+	}
 
 	.customization-toggle-wrapper {
-		padding: 1rem;
+		padding: 0.5rem;
+		// 2px to ensure the outline is visible,
+		// otherwise it is hidden behind header which has higher z-index
+		padding-top: 2px;
 
-		@include gb.break-medium() {
-			border-bottom: 1px solid #ddd;
-		}
-
-		:global(.components-base-control) {
+		:global(.components-toggle-group-control) {
+			max-height: 2rem;
 			max-width: 400px;
 			margin: 0 auto;
 		}
