--- projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/desktop/styles.module.scss
@@ -1,27 +1,0 @@
-@use "@automattic/jetpack-base-styles/gutenberg-base-styles" as gb;
-@use "../mixins.scss" as mixins;
-
-.tab-panel-desktop {
-
-	display: flex;
-	flex-direction: row;
-	height: 100%;
-
-	[role="tablist"] {
-
-		@include mixins.tab-list-overrides();
-		border-inline-end: 1px solid gb.$gray-200;
-	}
-
-	[role="tabpanel"] {
-		width: 100%;
-	}
-
-	@include mixins.tab-item-overrides();
-}
-
-.tab-content {
-	display: grid;
-	grid-template-columns: clamp(200px, 33.33%, 350px) minmax(0, 1fr);
-	height: 100%;
-}
