--- projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/mixins.scss
@@ -1,24 +1,0 @@
-@use "@automattic/jetpack-base-styles/gutenberg-base-styles" as gb;
-
-@mixin tab-list-overrides() {
-	padding: gb.$grid-unit-15; // 12px
-}
-
-@mixin tab-item-overrides() {
-
-	// Styles to override for .components-tab-panel__tabs-item
-	:global(.components-tab-panel__tabs-item) {
-		padding: 0;
-		height: 40px !important; // Core uses !important for tab height.
-
-		&:global(.is-active) {
-			background: gb.$gray-900;
-			border-radius: gb.$radius-medium;
-
-			// Hide the default active tab underline.
-			&::after {
-				display: none;
-			}
-		}
-	}
-}
