+++ projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/tab-content.tsx
@@ -1,7 +1,7 @@
 import { useSelect } from '@wordpress/data';
-import { store as socialStore } from '../../../../social-store';
-import { ConnectionToggle } from '../../connection-toggle';
-import { CustomizationSection } from '../../customization-section';
+import { store as socialStore } from '../../../social-store';
+import { ConnectionToggle } from '../connection-toggle';
+import { CustomizationSection } from '../customization-section';
 import { PreviewSection } from '../preview-section';
 import styles from './styles.module.scss';
 
@@ -24,7 +24,10 @@ export function TabContent( { connectionId, usingPerNetworkCustomization }: TabC
 	);
 
 	return (
-		<div className={ styles[ 'tab-content' ] }>
+		<div
+			className={ styles[ 'tab-content' ] }
+			data-variant={ usingPerNetworkCustomization ? 'per-network' : 'global' }
+		>
 			{ usingPerNetworkCustomization ? (
 				<CustomizationSection connection={ connection } usingPerNetworkCustomization />
 			) : (
